Implementing more handlers
